feat: Add flash check pipeline job
This PR sets up a way to add jobs to the casc via jobdsl.
It allows groovy scripts to be imported.
The first implementation is a flash check that flashes all hil nodes,
and if there is a failure then it reboots the node.
Note, it requires all pis to have RIOT with the binary for that board built in `opt/RIOT`

Another issue is that the ${} cannot be included in the .groovy script that easy since things get resolved by yml and jobdsl.
A workaround to this is adding
```
\$\
{my_variable}
```
instead of
```
${my_variable}
```