fix(nodes): make all nodes exclusive
This makes the nodes only run if a label is matched.
Since all jobs use matching labels anyways it should have not negative impact.
This only thing that will run without matching labels will be master.
This fixes the issue of using the seed job that depends on

pipeline from scm jenkinsscript

This was failing (maybe due to old git versions) when running on mobi15.
Forcing the seed job to run on master seems to be the best way.
There is also no way to specify that in the job itself.